requirements.txt
attrs==25.1.0
black==21.12b0
blinker==1.9.0
click==8.1.8
colorama==0.4.6
factory_boy==3.3.3
Faker==35.2.0
flake8==4.0.1
flake8-bugbear==22.1.11
flake8-pie==0.15.0
Flask==3.1.0
Flask-Cors==5.0.0
Flask-SQLAlchemy==3.1.1
gherkin-official==29.0.0
greenlet==3.1.1
iniconfig==2.0.0
isort==5.9.3
itsdangerous==2.2.0
Jinja2==3.1.5
Mako==1.3.9
MarkupSafe==3.0.2
marshmallow==3.26.1
mccabe==0.6.1
mypy==0.931
mypy-extensions==1.0.0
packaging==24.2
parse==1.20.2
parse_type==0.6.4
pathspec==0.12.1
platformdirs==4.3.6
pluggy==1.5.0
pycodestyle==2.8.0
pyflakes==2.4.0
pytest==8.3.4
pytest-bdd==8.1.0
python-dateutil==2.9.0.post0
six==1.17.0
SQLAlchemy==2.0.37
tomli==1.2.3
typing_extensions==4.12.2
Werkzeug==3.1.3
